refactor: Separate service with interface
