+++ src/main/java/com/woowacourse/lotto/controller/LottoController.java
@@ -3,7 +3,6 @@ package com.woowacourse.lotto.controller;
 import com.woowacourse.lotto.WebUILottoApplication;
 import com.woowacourse.lotto.domain.LottoBuyList;
 import com.woowacourse.lotto.domain.dto.LottoBuyDto;
-import com.woowacourse.lotto.service.LottoBuyService;
 import com.woowacourse.lotto.service.LottoBuyWebService;
 import spark.Request;
 import spark.Response;
@@ -12,7 +11,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class LottoController {
-    private static LottoBuyService lottoBuyService;
+    private static LottoBuyWebService lottoBuyService;
 
     public static Object init(Request request, Response response) {
         String moneyInput = request.queryParams("money");
